+++ lib/samsao/config.rb
@@ -1,6 +1,16 @@
 module Danger
   # Samsao's config class
   class SamsaoConfig
+    def initialize
+      @changelogs = ['CHANGELOG.md']
+    end
+
+    def changelogs(*entries)
+      return @changelogs if entries.nil? || entries.empty?
+
+      @changelogs = entries
+    end
+
     def sources(*entries)
       return @sources if entries.nil? || entries.empty?