+++ spec/samsao_spec.rb
@@ -1,3 +1,5 @@
+# rubocop:disable Metrics/ModuleLength
+
 require File.expand_path('../spec_helper', __FILE__)
 
 module Danger
@@ -31,7 +33,7 @@ module Danger
       end
 
       describe 'branching model' do
-        ['fix', 'feature', 'release', 'trivia'].each do |branch_prefix|
+        ['fix', 'feature', 'release', 'trivial'].each do |branch_prefix|
           it "continues on #{branch_prefix}/ prefix" do
             allow(@plugin.github).to receive(:branch_for_head).and_return("#{branch_prefix}/something")
 
@@ -46,7 +48,7 @@ module Danger
 
           @plugin.fail_when_wrong_branching_model
 
-          expect(@dangerfile).to have_error('Your branch should be prefixed with feature/, fix/, trivia/ or release/!')
+          expect(@dangerfile).to have_error('Your branch should be prefixed with feature/, fix/, trivial/ or release/!')
         end
 
         it 'fails on good prefix but wrong format' do
@@ -54,7 +56,7 @@ module Danger
 
           @plugin.fail_when_wrong_branching_model
 
-          expect(@dangerfile).to have_error('Your branch should be prefixed with feature/, fix/, trivia/ or release/!')
+          expect(@dangerfile).to have_error('Your branch should be prefixed with feature/, fix/, trivial/ or release/!')
         end
       end
 
@@ -85,6 +87,69 @@ module Danger
 
           expected_message = 'Your feature branch should have a single commit but found 2, squash them together!'
           expect(@dangerfile).to have_error(expected_message)
+        end
+      end
+
+      describe 'changelog updated' do
+        it 'continues on non-feature & non-fix branch' do
+          allow(@plugin.github).to receive(:branch_for_head).and_return('trivial/a')
+
+          @plugin.fail_when_changelog_update_missing
+
+          expect(@dangerfile).to have_no_error
+        end
+
+        it 'continues on feature branch and CHANGELOG updated' do
+          allow(@plugin.github).to receive(:branch_for_head).and_return('feature/a')
+          allow(@plugin.git).to receive(:modified_files).and_return(['CHANGELOG.md'])
+
+          @plugin.fail_when_changelog_update_missing
+
+          expect(@dangerfile).to have_no_error
+        end
+
+        it 'continues on fix branch and CHANGELOG updated' do
+          allow(@plugin.github).to receive(:branch_for_head).and_return('fix/a')
+          allow(@plugin.git).to receive(:modified_files).and_return(['CHANGELOG.md'])
+
+          @plugin.fail_when_changelog_update_missing
+
+          expect(@dangerfile).to have_no_error
+        end
+
+        it 'accepts customized changelogs path' do
+          @plugin.config do
+            changelogs 'CHANGELOG.yml'
+          end
+
+          allow(@plugin.github).to receive(:branch_for_head).and_return('fix/a')
+          allow(@plugin.git).to receive(:modified_files).and_return(['CHANGELOG.yml'])
+
+          @plugin.fail_when_changelog_update_missing
+
+          expect(@dangerfile).to have_no_error
+        end
+
+        it 'accepts multiple changelog paths' do
+          @plugin.config do
+            changelogs 'CHANGELOG.yml', 'web/CHANGELOG.md'
+          end
+
+          allow(@plugin.github).to receive(:branch_for_head).and_return('fix/a')
+          allow(@plugin.git).to receive(:modified_files).and_return(['web/CHANGELOG.md'])
+
+          @plugin.fail_when_changelog_update_missing
+
+          expect(@dangerfile).to have_no_error
+        end
+
+        it 'fails on feature branch and CHANGELOG not updated' do
+          allow(@plugin.github).to receive(:branch_for_head).and_return('feature/a')
+          allow(@plugin.git).to receive(:modified_files).and_return([])
+
+          @plugin.fail_when_changelog_update_missing
+
+          expect(@dangerfile).to have_error('You did a fix or a feature without updating CHANGELOG file!')
         end
       end
     end