+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## In progress
 
+ * Added `fail_when_changelog_update_missing` action.
+
  * Added `fail_when_non_single_commit_feature` action.